CricketIQ/requirements.txt
# ─── Core Python ────────────────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.2.0
polars>=0.20.0

# ─── Data Engineering ────────────────────────────────────────────────────────
duckdb>=0.10.0
dbt-core>=1.7.0
dbt-duckdb>=1.7.0
great-expectations>=0.18.0
prefect>=2.18.0
pyarrow>=15.0.0        # Parquet I/O
fastparquet>=2024.0.0  # Parquet alternative

# ─── ML & MLOps ──────────────────────────────────────────────────────────────
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.3.0
catboost>=1.2.0
optuna>=3.6.0
shap>=0.45.0
mlflow>=2.12.0
evidently>=0.4.0

# ─── API ─────────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
httpx>=0.27.0           # async HTTP client for tests

# ─── Dashboard ───────────────────────────────────────────────────────────────
streamlit>=1.33.0
plotly>=5.20.0
altair>=5.3.0

# ─── GenAI (optional) ────────────────────────────────────────────────────────
langchain>=0.1.0
langchain-openai>=0.1.0
chromadb>=0.4.0
openai>=1.20.0
tiktoken>=0.6.0

# ─── Testing & Code Quality ──────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=5.0.0
ruff>=0.4.0
mypy>=1.9.0
pre-commit>=3.7.0

# ─── Utilities ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pyyaml>=6.0.1
loguru>=0.7.2
rich>=13.7.0
tqdm>=4.66.0
click>=8.1.7
